+++ tests/api-tests/fields/unique.test.ts
@@ -19,6 +19,16 @@ multiAdapterRunners().map(({ runner, provider, after }) =>
       .forEach(mod => {
         (mod.testMatrix || ['default']).forEach((matrixValue: string) => {
           describe(`${mod.name} - ${matrixValue} - isUnique`, () => {
+            beforeEach(() => {
+              if (mod.beforeEach) {
+                mod.beforeEach();
+              }
+            });
+            afterEach(async () => {
+              if (mod.afterEach) {
+                await mod.afterEach();
+              }
+            });
             beforeAll(() => {
               if (mod.beforeAll) {
                 mod.beforeAll();
@@ -153,6 +163,7 @@ multiAdapterRunners().map(({ runner, provider, after }) =>
                       }),
                     }),
                     images: { upload: 'local', local: { storagePath: 'tmp_test_images' } },
+                    files: { upload: 'local', local: { storagePath: 'tmp_test_files' } },
                   }),
                 });
               } catch (error) {