+++ tests/api-tests/relationships/crud-self-ref/one-to-many-one-sided.test.ts
@@ -72,11 +72,8 @@ const createComplexData = async (context: KeystoneContext) => {
   expect(data2.createUsers[0].friend.name).toEqual('B1');
   expect(data2.createUsers[1].name).toEqual('C1');
 
-  type T = { allUsers: { id: IdType; name: string; friend: { id: IdType; name: string } }[] };
-  const { allUsers } = (await context.graphql.run({
-    query: '{ allUsers { id name friend { id name } } }',
-  })) as T;
-  return { users: allUsers };
+  const users = await context.lists.User.findMany({ query: 'id name friend { id name }' });
+  return { users };
 };
 
 const getUserAndFriend = async (context: KeystoneContext, userId: IdType, friendId: IdType) => {
@@ -122,10 +119,10 @@ multiAdapterRunners().map(({ runner, provider }) =>
                 ['D', 1],
                 ['E', 0],
               ].map(async ([name, count]) => {
-                const data = await context.graphql.run({
-                  query: `{ allUsers(where: { friend: { name_contains: "${name}"}}) { id }}`,
+                const users = await context.lists.User.findMany({
+                  where: { friend: { name_contains: name } },
                 });
-                expect(data.allUsers.length).toEqual(count);
+                expect(users.length).toEqual(count);
               })
             );
           })
@@ -134,20 +131,16 @@ multiAdapterRunners().map(({ runner, provider }) =>
           'is_null: true',
           runner(setupKeystone, async ({ context }) => {
             await createComplexData(context);
-            const data = await context.graphql.run({
-              query: `{ allUsers(where: { friend_is_null: true }) { id }}`,
-            });
-            expect(data.allUsers.length).toEqual(5);
+            const users = await context.lists.User.findMany({ where: { friend_is_null: true } });
+            expect(users.length).toEqual(5);
           })
         );
         test(
           'is_null: false',
           runner(setupKeystone, async ({ context }) => {
             await createComplexData(context);
-            const data = await context.graphql.run({
-              query: `{ allUsers(where: { friend_is_null: false }) { id }}`,
-            });
-            expect(data.allUsers.length).toEqual(4);
+            const users = await context.lists.User.findMany({ where: { friend_is_null: false } });
+            expect(users.length).toEqual(4);
           })
         );
       });
@@ -380,12 +373,11 @@ multiAdapterRunners().map(({ runner, provider }) =>
 
               // Check all the companies look how we expect
               await (async () => {
-                const data = await context.graphql.run({
-                  query: '{ allUsers(sortBy: name_ASC) { id name friend { id name } } }',
-                });
-                const users = data.allUsers.filter(
-                  ({ name }: { name: string }) => name.length === 1
-                );
+                const _users = (await context.lists.User.findMany({
+                  sortBy: ['name_ASC'],
+                  query: 'id name friend { id name }',
+                })) as { name: string; friend: { name: string } }[];
+                const users = _users.filter(({ name }: { name: string }) => name.length === 1);
                 const expected = [
                   ['A', 'A1'],
                   ['B', 'D1'],
@@ -410,12 +402,11 @@ multiAdapterRunners().map(({ runner, provider }) =>
 
               // Check all the friends look how we expect
               await (async () => {
-                const data = await context.graphql.run({
-                  query: '{ allUsers(sortBy: name_ASC) { id name } }',
-                });
-                const friends = data.allUsers.filter(
-                  ({ name }: { name: string }) => name.length === 2
-                );
+                const _users = (await context.lists.User.findMany({
+                  sortBy: ['name_ASC'],
+                  query: 'id name',
+                })) as { name: string }[];
+                const friends = _users.filter(({ name }: { name: string }) => name.length === 2);
                 expect(friends[0].name).toEqual('A1');
                 expect(friends[1].name).toEqual('B1');
                 expect(friends[2].name).toEqual('C1');
@@ -441,12 +432,11 @@ multiAdapterRunners().map(({ runner, provider }) =>
 
               // Check all the companies look how we expect
               await (async () => {
-                const data = await context.graphql.run({
-                  query: '{ allUsers(sortBy: name_ASC) { id name friend { id name } } }',
-                });
-                const users = data.allUsers.filter(
-                  ({ name }: { name: string }) => name.length === 1
-                );
+                const _users = (await context.lists.User.findMany({
+                  sortBy: ['name_ASC'],
+                  query: 'id name friend { id name }',
+                })) as { name: string; friend: { name: string } }[];
+                const users = _users.filter(({ name }) => name.length === 1);
                 expect(users[0].name).toEqual('A');
                 if (name === 'A1') {
                   expect(users[0].friend).toBe(null);
@@ -477,12 +467,11 @@ multiAdapterRunners().map(({ runner, provider }) =>
 
               // Check all the friends look how we expect
               await (async () => {
-                const data = await context.graphql.run({
-                  query: '{ allUsers(sortBy: name_ASC) { id name } }',
-                });
-                const friends = data.allUsers.filter(
-                  ({ name }: { name: string }) => name.length === 2
-                );
+                const _users = (await context.lists.User.findMany({
+                  sortBy: ['name_ASC'],
+                  query: 'id name',
+                })) as { name: string }[];
+                const friends = _users.filter(({ name }) => name.length === 2);
                 const expected = ['A1', 'B1', 'C1', 'D1'].filter(x => x !== name);
                 expect(friends[0].name).toEqual(expected[0]);
                 expect(friends[1].name).toEqual(expected[1]);